refactor: move homepage client component code into a seperate component
